requirements.txt
pyasn1
pyasn1-modules
python-dotenv
cryptography
boto3